acceptance: consolidate language tests
These language tests are now nothing but wrappers around a testdata
script, so they don't need their own files.

Release note: None
